// src/enums/assets.ts
export enum AssetGetByEntityInclude {
  QR_CODE = 'QR_CODE',
  FEEDS_TO = 'FEEDS_TO',
  FED_FROM = 'FED_FROM',
  PROPS = 'PROPS',
  PURCHASES = 'PURCHASES',
  NOTES = 'NOTES',
  FILES = 'FILES',
  TYPE = 'TYPE',
  CATEGORY = 'CATEGORY',
  ROOM = 'ROOM',
  FLOOR = 'FLOOR',
  BUILDING = 'BUILDING',
  REGION = 'REGION',
  WORK_ORDER = 'WORK_ORDER',
  CREATOR = 'CREATOR',
  AVATAR = 'AVATAR',
}

export enum AssetGetByEntityAttributes {
  SERIAL_NUMBER = 'SERIAL_NUMBER',
  EQUIPMENT_ID = 'EQUIPMENT_ID',
  MODEL = 'MODEL',
  IS_CRITICAL = 'IS_CRITICAL',
  LABOR_VALUE = 'LABOR_VALUE',
  COST = 'COST',
  IS_NEW = 'IS_NEW',
  MANUFACTURER = 'MANUFACTURER',
  INSTALL_DATE = 'INSTALL_DATE',
  ASSET_AGE = 'ASSET_AGE',
  DESCRIPTION = 'DESCRIPTION',
  PAGES_COUNT = 'PAGES_COUNT',
  EMERGENCY_PLANS_COUNT = 'EMERGENCY_PLANS_COUNT',
  HAS_PLANS = 'HAS_PLANS',
  HAS_EMERGENCY_PLANS = 'HAS_EMERGENCY_PLANS',
}

export enum EverythingForkEnum {
  CATEGORY = 'CATEGORY',
  TYPE = 'TYPE',
  ASSET_AGE = 'ASSET_AGE',
  ASSET_COST = 'ASSET_COST',
}

export enum InventoryItemStatus {
  AVAILABLE = 'Available',
  UNAVAILABLE = 'UNAVAILABLE',
  ALLOCATED_FOR_WO = 'Allocated for WO',
  ON_HOLD = 'On Hold',
  TRANSFER_REQUESTED = 'Transfer Requested',
}